Added Player Stats Code
From different player class and removed it.
