--- Assets/Player.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bcd6900b5b5e8934eb163ede8bb6160b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
